requirements.txt
Django==4.2.20
pytz==2021.1
sqlparse==0.4.1
asgiref==3.8.1
django-ckeditor==6.7.2
django-js-asset==2.2.0
typing-extensions==4.12.2
tzdata==2025.2
